# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {},
	"info": {
		"title": "rsclone-be",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
		"/statistics": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"additionalProperties": true
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"StatisticsController"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"tags": [
					"StatisticsController"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"userId": {
									"type": "string"
								},
								"score": {
									"type": "string"
								},
								"totalTime": {
									"type": "string"
								}
							},
							"title": "CreateBody",
							"type": "object",
							"required": [
								"userId",
								"score",
								"totalTime"
							]
						}
					}
				]
			}
		},
		"/statistics/{userId}": {
			"get": {
				"operationId": "GetByUserId",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"additionalProperties": true
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"StatisticsController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "DeleteByUserId",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"StatisticsController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"userName": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							},
							"title": "CreateBody",
							"type": "object",
							"required": [
								"userName",
								"password"
							]
						}
					}
				]
			}
		},
		"/auth/test": {
			"post": {
				"operationId": "Test",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"userName": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							},
							"title": "TestBody",
							"type": "object",
							"required": [
								"userName",
								"password"
							]
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"userName": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							},
							"title": "RegisterBody",
							"type": "object",
							"required": [
								"userName",
								"password"
							]
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}